+++ src/tables.jl
@@ -1,13 +1,15 @@
 #-----------------------------------------------------------------------# IndexedTable
-Tables.istable(::Type{IndexedTable{C}}) where {C<:Columns} = Tables.istable(C)
-Tables.materializer(t::IndexedTable) = table
-for f in [:rowaccess, :rows, :columnaccess, :columns, :schema]
-    @eval Tables.$f(t::IndexedTable) = Tables.$f(Columns(columns(t)))
-end
+Tables.istable(::Type{<:IndexedTable}) = true
 
-#-----------------------------------------------------------------------# NDSparse
-# Tables.istable(::Type{NDSparse{T,D,C,V}}) where {T,D,C<:TableColumns,V<:TableColumns} = true
-# Tables.materializer(t::NDSparse) = ndpsarse
+Tables.materializer(t::IndexedTable) = table
+
+Tables.columnaccess(::Type{<:IndexedTable}) = Tables.columnaccess(StructArray)
+Tables.columns(t::IndexedTable) = Tables.columns(columns(t))
+
+Tables.rowaccess(::Type{<:IndexedTable}) = Tables.rowaccess(StructArray)
+Tables.rows(t::IndexedTable) = Tables.rows(rows(t))
+
+# table(x; copy=false, kw...) = table(Tables.columntable(x); copy=copy, kw...)