+++ test/runtests.jl
@@ -1,6 +1,6 @@
 using Test, IndexedTables, OnlineStats, WeakRefStrings, Tables, Random, Dates, 
     PooledArrays, SparseArrays, WeakRefStrings, LinearAlgebra, Statistics,
-    TableTraits, IteratorInterfaceExtensions, Serialization
+    TableTraits, IteratorInterfaceExtensions, Serialization, DataValues
 
 using IndexedTables: excludecols, sortpermby, primaryperm, best_perm_estimate, hascolumns,
     collect_columns_flattened